+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Uses 'mirai' to run `oanda_studio(new.process = TRUE)`
 * Fixes `oanda_switch()`, regression since v1.4.11.
+* Updates internal OANDA instruments list.
 * Internal performance enhancements.
 
 # ichimoku 1.4.11